# Aufgabe 6.2/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Library API",
    "description": "This is an API for managing books and lends in a library.",
    "version": "1.0.0"
  },
  "paths": {
    "/books": {
      "get": {
        "summary": "Get a list of all books",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new book",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewBook"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      }
    },
    "/books/{isbn}": {
      "put": {
        "summary": "Update a book by ISBN",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBook"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a book by ISBN",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/lends": {
      "get": {
        "summary": "Get a list of all lends",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lend"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new lend",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewLend"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lend"
                }
              }
            }
          }
        }
      }
    },
    "/lends/{isbn}": {
      "get": {
        "summary": "Get a lend by ISBN",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lend"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a lend by ISBN",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLend"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lend"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a lend by ISBN",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          }
        }
      },
      "NewBook": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          }
        }
      },
      "UpdateBook": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          }
        }
      },
      "Lend": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "NewLend": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "UpdateLend": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "string",
            "format": "date"
          }
        }
      }
    }
  }
}